DRILL-7362: COUNT(*) on JSON with outer list results in JsonParse error
